+++ src/onegov.core/onegov/core/orm/abstract/__init__.py
@@ -1,3 +1,4 @@
+from onegov.core.orm.abstract.associable import Associable, associated
 from onegov.core.orm.abstract.adjacency_list import (
     AdjacencyList, AdjacencyListCollection, MoveDirection, sort_siblings
 )
@@ -5,6 +6,8 @@ from onegov.core.orm.abstract.adjacency_list import (
 __all__ = [
     'AdjacencyList',
     'AdjacencyListCollection',
+    'Associable',
+    'associated',
     'MoveDirection',
     'sort_siblings'
 ]